+++ cmd/agyent/update.go
@@ -102,4 +102,3 @@ func init() {
 
 	rootCmd.AddCommand(updateCmd)
 }
-